Move length control file
